rename some file and navigate learn page from learning page
