+++ voice_bridge/lib/screens/sessionScreen/controllers/sessionController.dart
@@ -5,7 +5,7 @@ import 'package:get/get.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 import '../../../resources/routes/routesName.dart';
-import '../view_model/session_model.dart';
+import '../view_model/sessionModel.dart';
 // import 'package:cloud_firestore/cloud_firestore.dart';
 
 class SessionController extends GetxController {
