+++ voice_bridge/lib/screens/sessionScreen/view_model/lessonModel.dart
