+++ voice_bridge/lib/screens/sessionScreen/view_model/sessionModel.dart
@@ -1,4 +1,4 @@
-import 'lesson_model.dart';
+import 'lessonModel.dart';
 
 class Session {
   final String sessionName;
